#87 Fix expansion for norminette
