+++ includes/minishell.h
@@ -121,10 +121,12 @@ int				is_separating_word(char *line, int pos);
 int				is_quote(char c);
 
 t_token			*make_token(char *line, size_t pos, size_t len, t_token_kind a);
+void			ft_lstdelone_all(t_list *lst, void (*del)(void*));
+void			ft_lstclear_all(t_list **lst, void (*del)(void*));
+
 int				ft_lstadd_token(t_list **token_list, t_token *new_token);
 int				ft_lstadd_word(t_list **lst, char *new_word);
 void			ft_lstadd_last(t_list **lst, t_list *new);
-void			ft_lstclear_all(t_list **lst, void (*del)(void*));
 
 void			do_piping(int pipes[2], t_node *node, t_global_state *state);
 void			close_pipes(int pipes[2], t_node *node, t_global_state *state);
@@ -157,7 +159,6 @@ t_list			*check_quote(t_expand_state *e_state);
 int				expand(t_list *token_lst, t_list **expanded_lst, t_envs *envs,
 					int exit_status);
 void			init_expand_state(t_expand_state *e_state);
-void			ft_lstdelone_all(t_list *lst, void (*del)(void*));
 
 t_node			*parse(t_list **token_list);
 int				is_command_token(t_list **token_list);
